requirements.txt
discord.py[voice]
wolframalpha
requests
gTTS